+++ tests/testthat/test-2-worker.r
@@ -1,77 +1,90 @@
 context("worker usage")
 
-test_that("timeouts are triggered correctly", {
-    m = methods::new(CMQMaster)
-    addr = m$listen("inproc://endpoint")
-    expect_error(m$recv(0L))
-    m$close(0L)
-
+test_that("connect to invalid endpoint errors", {
     w = methods::new(CMQWorker)
     expect_error(w$connect("tcp://localhost:12345", 0L))
     w$close()
+})
+
+test_that("recv without pending workers errors before timeout", {
+    m = methods::new(CMQMaster)
+    addr = m$listen("inproc://endpoint")
+    expect_error(m$recv(-1L))
+    m$close(500L)
+})
+
+test_that("recv timeout works", {
+    m = methods::new(CMQMaster)
+    addr = m$listen("inproc://endpoint")
+    m$add_pending_workers(1L)
+    expect_error(m$recv(0L))
+    m$close(500L)
 })
 
 test_that("worker evaluation", {
     m = methods::new(CMQMaster)
     w = methods::new(CMQWorker, m$context())
     addr = m$listen("inproc://endpoint")
-    w$connect(addr, 0L)
+    m$add_pending_workers(1L)
+    w$connect(addr, 500L)
 
-    m$recv(0L)
+    m$recv(500L)
     m$send(expression(5 * 2))
     status = w$process_one()
-    result = m$recv(0L)
+    result = m$recv(500L)
 
     expect_true(status)
     expect_equal(result, 10)
 
     w$close()
-    m$close(0L)
+    m$close(500L)
 })
 
 test_that("export variable to worker", {
     m = methods::new(CMQMaster)
     w = methods::new(CMQWorker, m$context())
     addr = m$listen("inproc://endpoint")
-    w$connect(addr, 0L)
+    m$add_pending_workers(1L)
+    w$connect(addr, 500L)
 
     m$add_env("x", 3)
-    m$recv(0L)
+    m$recv(500L)
     m$send(expression(5 + x))
     status = w$process_one()
-    result = m$recv(0L)
+    result = m$recv(500L)
     expect_true(status)
     expect_equal(result, 8)
 
     m$add_env("x", 5)
     m$send(expression(5 + x))
     status = w$process_one()
-    result = m$recv(0L)
+    result = m$recv(500L)
     expect_true(status)
     expect_equal(result, 10)
 
     w$close()
-    m$close(0L)
+    m$close(500L)
 })
 
 test_that("load package on worker", {
     m = methods::new(CMQMaster)
     w = methods::new(CMQWorker, m$context())
     addr = m$listen("inproc://endpoint")
-    w$connect(addr, 0L)
+    m$add_pending_workers(1L)
+    w$connect(addr, 500L)
 
     m$add_pkg("parallel")
 
-    m$recv(0L)
+    m$recv(500L)
     m$send(expression(splitIndices(1, 1)[[1]]))
     status = w$process_one()
-    result = m$recv(0L)
+    result = m$recv(500L)
 
     expect_true(status)
     expect_equal(result, 1)
 
     w$close()
-    m$close(0L)
+    m$close(500L)
 })
 
 test_that("errors are sent back to master", {
@@ -80,18 +93,19 @@ test_that("errors are sent back to master", {
     m = methods::new(CMQMaster)
     w = methods::new(CMQWorker, m$context())
     addr = m$listen("inproc://endpoint")
-    w$connect(addr, 0L)
+    m$add_pending_workers(1L)
+    w$connect(addr, 500L)
 
-    m$recv(0L)
+    m$recv(500L)
     m$send(expression(stop("errmsg")))
     status = w$process_one()
-    result = m$recv(0L)
+    result = m$recv(500L)
 
     expect_true(status)
     expect_true(inherits(result, c("condition", "worker_error")))
 
     w$close()
-    m$close(0L)
+    m$close(500L)
 })
 
 test_that("worker R API", {
@@ -100,6 +114,7 @@ test_that("worker R API", {
 
     m = methods::new(CMQMaster)
     addr = m$listen("tcp://127.0.0.1:*")
+    m$add_pending_workers(1L)
 #    addr = m$listen("inproc://endpoint") # mailbox.cpp assertion error
 
     p = parallel::mcparallel(worker(addr))
@@ -111,7 +126,7 @@ test_that("worker R API", {
     m$send_shutdown()
     pc = parallel::mccollect(p, wait=TRUE, timeout=0.5)
     expect_equal(pc[[1]], NULL)
-    m$close(0L)
+    m$close(500L)
 })
 
 test_that("communication with two workers", {
@@ -120,12 +135,13 @@ test_that("communication with two workers", {
 
     m = methods::new(CMQMaster)
     addr = m$listen("tcp://127.0.0.1:*")
+    m$add_pending_workers(2L)
     w1 = parallel::mcparallel(worker(addr))
     w2 = parallel::mcparallel(worker(addr))
 
     expect_null(m$recv(1000L)) # worker 1 up
     m$send(expression({ Sys.sleep(0.5); 5 + 2 }))
-    expect_null(m$recv(1000L)) # worker 2 up
+    expect_null(m$recv(500L)) # worker 2 up
     m$send(expression({ Sys.sleep(0.5); 3 + 1 }))
     r1 = m$recv(1000L)
     m$send_shutdown()
@@ -138,5 +154,5 @@ test_that("communication with two workers", {
     coll2 = parallel::mccollect(w2, wait=TRUE, timeout=0.5)
     expect_equal(names(coll2), as.character(w2$pid))
 
-    m$close(0L)
+    m$close(500L)
 })