+++ tests/testthat/test-4-pool.r
@@ -4,10 +4,26 @@ test_that("starting and stopping multicore", {
     skip_on_os("windows")
 
     w = workers(1, qsys_id="multicore")
+    expect_equal(w$workers_total, 1)
+    expect_equal(w$workers_running, 0)
     expect_null(w$recv())
+    expect_equal(w$workers_running, 1)
     w$send(3 + 4)
+    expect_equal(w$workers_running, 1)
     expect_equal(w$recv(), 7)
+    expect_equal(w$workers_running, 1)
     w$send_shutdown()
+    expect_equal(w$workers_running, 0)
+    expect_equal(w$workers_total, 1)
+    expect_error(w$send(1))
+    expect_error(w$recv())
+    w$cleanup()
+    expect_equal(w$workers_running, 0)
+    expect_equal(w$workers_total, 0)
+    expect_error(w$send(2))
+    expect_error(w$recv())
+    expect_equal(w$workers_running, 0)
+    expect_equal(w$workers_total, 0)
 })
 
 test_that("calculations are really done on the worker", {
@@ -20,6 +36,7 @@ test_that("calculations are really done on the worker", {
     w$send(x + y, x=4)
     expect_equal(w$recv(), 7)
     w$send_shutdown()
+    w$cleanup()
 })
 
 test_that("multiprocess", {
@@ -30,6 +47,7 @@ test_that("multiprocess", {
     w$send(3 + 5)
     expect_equal(w$recv(), 8)
     w$send_shutdown()
+    w$cleanup()
 })
 
 test_that("work_chunk on multiprocess", {
@@ -43,4 +61,24 @@ test_that("work_chunk on multiprocess", {
     expect_equal(res$warnings, list())
     expect_equal(res$errors, list())
     w$send_shutdown()
+    w$cleanup()
+})
+
+test_that("worker creation passes template filling values", {
+    skip("works interactively")
+    TMPL_FILLER = R6::R6Class("TMPL_FILLER",
+        inherit = QSys,
+        public = list(
+            initialize = function(addr, n_jobs, master, ...) {
+                super$initialize(addr=addr, master=master, template="LSF")
+                self$filled = private$fill_options(...)
+            },
+            filled = list()
+        )
+    )
+    options(clustermq.defaults = list(cores="defaults_test", memory="invalid"))
+    w = workers(1, qsys_id="tmpl_filler", template=list(memory="test"))
+
+    expect_equal(w$workers$filled$memory, "test")
+    expect_equal(w$workers$filled$cores, "defaults_test")
 })