+++ tests/testthat/test-7-ssh_proxy.r
@@ -6,20 +6,23 @@ has_localhost = has_connectivity("127.0.0.1")
 # the first test twice leads to a segfault; not sure why, fix this eventually
 test_that("simple forwarding works", {
     skip_if_not(has_localhost)
+    skip_on_cran()
+
     m = methods::new(CMQMaster)
     p = methods::new(CMQProxy)#, m$context())
     w = methods::new(CMQWorker)#, m$context())
     addr1 = m$listen("tcp://127.0.0.1:*")#"inproc://master")
     addr2 = p$listen("tcp://127.0.0.1:*")#"inproc://proxy")
+    m$add_pending_workers(1L)
     p$connect(addr1, 500L)
     w$connect(addr2, 500L)
     expect_true(p$process_one())
-    expect_null(m$recv(0L)) # worker up
+    expect_null(m$recv(500L)) # worker up
     m$send(5 + 2)
     expect_true(p$process_one())
     expect_true(w$process_one())
     expect_true(p$process_one())
-    result = m$recv(0L)
+    result = m$recv(500L)
     expect_equal(result, 7)
 
     w$close()
@@ -29,6 +32,7 @@ test_that("simple forwarding works", {
 
 test_that("proxy communication yields submit args", {
     skip_if_not(has_localhost)
+    skip_on_cran()
 
     m = methods::new(CMQMaster)
     p = methods::new(CMQProxy)#, m$context())
@@ -58,8 +62,8 @@ test_that("using the proxy without pool and forward", {
     addr = m$listen("tcp://127.0.0.1:*")
     p = parallel::mcparallel(ssh_proxy(sub(".*:", "", addr)))
 
-    m$proxy_submit_cmd(list(n_jobs=1), 500L)
-    expect_null(m$recv(500L)) # worker 1 up
+    m$proxy_submit_cmd(list(n_jobs=1), 1000L)
+    expect_null(m$recv(1000L)) # worker 1 up
     m$send(5 + 2)
     expect_equal(m$recv(500L), 7) # collect results
 
@@ -80,8 +84,8 @@ test_that("using the proxy without pool and forward, 2 workers", {
     addr = m$listen("tcp://127.0.0.1:*")
     p = parallel::mcparallel(ssh_proxy(sub(".*:", "", addr)))
 
-    m$proxy_submit_cmd(list(n_jobs=2), 500L)
-    expect_null(m$recv(500L)) # worker 1 up
+    m$proxy_submit_cmd(list(n_jobs=2), 1000L)
+    expect_null(m$recv(1000L)) # worker 1 up
     m$send({ Sys.sleep(0.5); 5 + 2 })
     expect_null(m$recv(500L)) # worker 2 up
     m$send({ Sys.sleep(0.5); 3 + 1 })
@@ -110,7 +114,7 @@ test_that("full SSH connection", {
             message="options(clustermq.scheduler') must be 'MULTICORE'")
 
     options(clustermq.template = "SSH", clustermq.ssh.host="127.0.0.1")
-    w = workers(n_jobs=1, qsys_id="ssh", reuse=FALSE, log_worker=T)
+    w = workers(n_jobs=1, qsys_id="ssh", reuse=FALSE)
     result = Q(identity, 42, n_jobs=1, timeout=10L, workers=w)
     expect_equal(result, list(42))
 })